Game database on mongo is now working perfectly
Created a mongoDB database called "Games". This contains the game we have inserted and the class MongoDBGame handles all the CRUD functionalities that are needed to create, read, update and delete gams in the database. A example of how this works is made in the program.cs in the Console App for testing only project.